requirements.txt
asgiref==3.11.1
Django==6.0.3
django-modeltranslation==0.20.2
pillow==12.1.1
sqlparse==0.5.5
typing_extensions==4.15.0


